requirements.txt
opencv-python
numpy
dash
plotly
mediapipe
IPython
pretty_midi
music21
musicalbeeps
pygame